feat: Add --pull flag to pull changes before commit
This commit introduces a new `--pull` flag to the CLI. When this flag is used, the tool will attempt to pull changes from the remote repository before proceeding with the commit. If the pull fails, an error message will be displayed, and the commit process will be aborted. This ensures that the local branch is up-to-date before making a commit.